# api-docs.json
{"openapi":"3.1.0","info":{"title":"School ERP MVP API","description":"API documentation for the School ERP Data Accessor.","contact":{"name":"Capscrum Development Team"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Role Management","description":"View roles and permissions (Super Admin only)"},{"name":"Authentication","description":"Login, register, and user profile endpoints"},{"name":"User Management","description":"CRUD operations for users (Admin only)"}],"paths":{"/api/users/{id}":{"get":{"tags":["User Management"],"summary":"Get user by ID","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"put":{"tags":["User Management"],"summary":"Update user profile","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"delete":{"tags":["User Management"],"summary":"Delete user","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users/{id}/roles":{"put":{"tags":["User Management"],"summary":"Assign roles to a user","description":"Replaces all current roles with the provided set of role names","operationId":"assignRoles","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"},"uniqueItems":true}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/subjects/{id}":{"get":{"tags":["subject-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Subject"}}}}}},"put":{"tags":["subject-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subject"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Subject"}}}}}},"delete":{"tags":["subject-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/students/{id}":{"get":{"tags":["student-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}},"put":{"tags":["student-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}},"delete":{"tags":["student-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/classes/{id}":{"get":{"tags":["school-class-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchoolClass"}}}}}},"put":{"tags":["school-class-controller"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolClass"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchoolClass"}}}}}},"delete":{"tags":["school-class-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/class-subjects/{id}":{"get":{"tags":["class-subject-controller"],"operationId":"getById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClassSubject"}}}}}},"put":{"tags":["class-subject-controller"],"operationId":"update_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassSubject"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClassSubject"}}}}}},"delete":{"tags":["class-subject-controller"],"operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/admissions/{id}":{"get":{"tags":["admission-controller"],"operationId":"getById_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Admission"}}}}}},"put":{"tags":["admission-controller"],"operationId":"update_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Admission"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Admission"}}}}}},"delete":{"tags":["admission-controller"],"operationId":"delete_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/academic-years/{id}":{"get":{"tags":["academic-year-controller"],"operationId":"getById_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AcademicYear"}}}}}},"put":{"tags":["academic-year-controller"],"operationId":"update_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AcademicYear"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AcademicYear"}}}}}},"delete":{"tags":["academic-year-controller"],"operationId":"delete_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/subjects":{"get":{"tags":["subject-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Subject"}}}}}}},"post":{"tags":["subject-controller"],"summary":"Create a new item","description":"This replaces the existing item if it already exists","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subject"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Subject"}}}}}}},"/api/students":{"get":{"tags":["student-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Student"}}}}}}},"post":{"tags":["student-controller"],"summary":"Create a new item","description":"This replaces the existing item if it already exists","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}}},"/api/classes":{"get":{"tags":["school-class-controller"],"operationId":"getAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SchoolClass"}}}}}}},"post":{"tags":["school-class-controller"],"summary":"Create a new item","description":"This replaces the existing item if it already exists","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolClass"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SchoolClass"}}}}}}},"/api/class-subjects":{"get":{"tags":["class-subject-controller"],"operationId":"getAll_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClassSubject"}}}}}}},"post":{"tags":["class-subject-controller"],"summary":"Create a new item","description":"This replaces the existing item if it already exists","operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassSubject"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClassSubject"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new user","description":"Creates a new user account and returns a JWT access token","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login with username/email and password","description":"Returns a JWT access token on successful authentication","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/admissions":{"get":{"tags":["admission-controller"],"operationId":"getAll_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Admission"}}}}}}},"post":{"tags":["admission-controller"],"summary":"Create a new item","description":"This replaces the existing item if it already exists","operationId":"create_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Admission"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Admission"}}}}}}},"/api/admissions/{id}/reject":{"post":{"tags":["admission-controller"],"summary":"Reject an admission application","operationId":"rejectAdmission","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"reason","in":"query","required":true,"schema":{"type":"string"}},{"name":"reviewedBy","in":"query","required":false,"schema":{"type":"string","default":"system"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Admission"}}}}}}},"/api/admissions/{id}/approve":{"post":{"tags":["admission-controller"],"summary":"Approve an admission and convert to Student","operationId":"approveAdmission","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"reviewedBy","in":"query","required":false,"schema":{"type":"string","default":"system"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Admission"}}}}}}},"/api/academic-years":{"get":{"tags":["academic-year-controller"],"operationId":"getAll_5","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AcademicYear"}}}}}}},"post":{"tags":["academic-year-controller"],"summary":"Create a new item","description":"This replaces the existing item if it already exists","operationId":"create_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AcademicYear"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AcademicYear"}}}}}}},"/api/users":{"get":{"tags":["User Management"],"summary":"Get all users","operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"/api/roles":{"get":{"tags":["Role Management"],"summary":"Get all roles","operationId":"getAllRoles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Role"}}}}}}}},"/api/roles/{id}":{"get":{"tags":["Role Management"],"summary":"Get role by ID with permissions","operationId":"getRoleById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Role"}}}}}}},"/api/roles/name/{name}":{"get":{"tags":["Role Management"],"summary":"Get role by name","operationId":"getRoleByName","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Role"}}}}}}},"/api/auth/me":{"get":{"tags":["Authentication"],"summary":"Get current user profile","description":"Returns the profile of the currently authenticated user","operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"components":{"schemas":{"UserDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"},"enabled":{"type":"boolean"},"roles":{"type":"array","items":{"type":"string"},"uniqueItems":true},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"School":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"string"},"schoolCode":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"}}},"Subject":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"string"},"school":{"$ref":"#/components/schemas/School"},"subjectCode":{"type":"string"},"name":{"type":"string"},"type":{"type":"string"},"isElective":{"type":"boolean"}}},"Student":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"string"},"school":{"$ref":"#/components/schemas/School"},"studentCode":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"gender":{"type":"string"},"bloodGroup":{"type":"string"},"photoUrl":{"type":"string"},"address":{"type":"string"},"previousSchool":{"type":"string"},"medicalNotes":{"type":"string"},"status":{"type":"string"}}},"AcademicYear":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"string"},"school":{"$ref":"#/components/schemas/School"},"label":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"isCurrent":{"type":"boolean"}}},"SchoolClass":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"string"},"school":{"$ref":"#/components/schemas/School"},"academicYear":{"$ref":"#/components/schemas/AcademicYear"},"classCode":{"type":"string"},"grade":{"type":"string"},"section":{"type":"string"},"roomNumber":{"type":"string"},"capacity":{"type":"integer","format":"int32"}}},"ClassSubject":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"string"},"schoolClass":{"$ref":"#/components/schemas/SchoolClass"},"subject":{"$ref":"#/components/schemas/Subject"}}},"Admission":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"deletedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string"},"updatedBy":{"type":"string"},"id":{"type":"string"},"school":{"$ref":"#/components/schemas/School"},"admissionCode":{"type":"string"},"applicantName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"gender":{"type":"string"},"gradeApplyingFor":{"type":"string"},"academicYearId":{"type":"string"},"parentName":{"type":"string"},"parentPhone":{"type":"string"},"parentEmail":{"type":"string"},"previousSchool":{"type":"string"},"status":{"type":"string"},"rejectionReason":{"type":"string"},"student":{"$ref":"#/components/schemas/Student"}}},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":50,"minLength":3},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","maxLength":100,"minLength":6},"fullName":{"type":"string","minLength":1},"roleNames":{"type":"array","items":{"type":"string"},"uniqueItems":true}},"required":["email","fullName","password","username"]},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"tokenType":{"type":"string"},"username":{"type":"string"},"roles":{"type":"array","items":{"type":"string"},"uniqueItems":true}}},"LoginRequest":{"type":"object","properties":{"usernameOrEmail":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","usernameOrEmail"]},"Permission":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"module":{"type":"string"}}},"Role":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"permissions":{"type":"array","items":{"$ref":"#/components/schemas/Permission"},"uniqueItems":true}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description — Please input your JWT below.","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}