# src/main/java/com/capscrum/school/erp/dataaccessor/dto/LoginRequest.java
package com.capscrum.school.erp.dataaccessor.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class LoginRequest {

    @NotBlank(message = "Username or email is required")
    private String usernameOrEmail;

    @NotBlank(message = "Password is required")
    private String password;
}