# src/main/java/com/capscrum/school/erp/dataaccessor/model/Permission.java
package com.capscrum.school.erp.dataaccessor.model;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Entity
@EqualsAndHashCode
@Table(name = "permissions")
public class Permission {
    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    private String id;
    
    @Column(unique = true, nullable = false)
    private String name;          // e.g. "STUDENT_READ", "ADMISSION_WRITE"
    
    private String description;
    
    @Column(nullable = false)
    private String module;        // e.g. "ADMISSION", "STUDENT", "FEE"
}